organize file systems and split up tetrimino logic
