+++ typescript-games/tetris/src/tetriminos/index.ts
@@ -1,0 +1,3 @@
+import { ALL_TETRIMINOS, Tetrimino } from "./pieces";
+
+export { ALL_TETRIMINOS, Tetrimino };
