+++ typescript-games/tetris/src/tetriminos/pieces.ts
@@ -1,4 +1,5 @@
-import { GridCoordinate } from "./grid";
+import { GridCoordinate } from "../grid";
+import { IPieceRotationStrategy, OPieceRotationStrategy, RotationStrategy, TSZJLPieceRotationStrategy } from "./rotation";
 
 enum ROTATION {
   NORTH,
@@ -154,65 +155,3 @@ export const ALL_TETRIMINOS = [
   TPiece,
   ZPiece,
 ]
-
-
-abstract class RotationStrategy {
-  piece: Tetrimino;
-
-  constructor(piece: Tetrimino) {
-    this.piece = piece
-  }
-
-  abstract rotateClockwise()
-
-  abstract rotateCounterClockwise()
-}
-
-
-class OPieceRotationStrategy extends RotationStrategy {
-  rotateClockwise() {
-
-  }
-
-  rotateCounterClockwise() {
-
-  }
-}
-
-
-class IPieceRotationStrategy extends RotationStrategy {
-  rotateClockwise() {
-
-  }
-
-  rotateCounterClockwise() {
-
-  }
-}
-
-
-class TSZJLPieceRotationStrategy extends RotationStrategy {
-  rotateCounterClockwise() {
-    const rotatedCoordinates = this.piece.coords.map(
-      coord => {
-        const transposedCoord = new GridCoordinate({ col: coord.col - this.piece.pivot.col, row: coord.row - this.piece.pivot.row });
-        const rotatedCoord = new GridCoordinate({ col: transposedCoord.row, row: - transposedCoord.col });
-        return new GridCoordinate({ col: rotatedCoord.col + this.piece.pivot.col, row: rotatedCoord.row + this.piece.pivot.row })
-      });
-
-    // go through all 5 transformations to see
-    this.piece.coords = rotatedCoordinates;
-  }
-
-  rotateClockwise() {
-    const rotatedCoordinates = this.piece.coords.map(
-      coord => {
-        const transposedCoord = new GridCoordinate({ col: coord.col - this.piece.pivot.col, row: coord.row - this.piece.pivot.row });
-        const rotatedCoord = new GridCoordinate({ col: - transposedCoord.row, row: transposedCoord.col });
-        return new GridCoordinate({ col: rotatedCoord.col + this.piece.pivot.col, row: rotatedCoord.row + this.piece.pivot.row })
-      });
-
-    // go through all 5 transformations to see
-    this.piece.coords = rotatedCoordinates;
-  }
-}
