+++ typescript-games/tetris/src/tetriminos/rotation.ts
@@ -1,0 +1,63 @@
+import { GridCoordinate } from "../grid";
+import { Tetrimino } from "./pieces";
+
+export abstract class RotationStrategy {
+  piece: Tetrimino;
+
+  constructor(piece: Tetrimino) {
+    this.piece = piece
+  }
+
+  abstract rotateClockwise()
+
+  abstract rotateCounterClockwise()
+}
+
+
+export class OPieceRotationStrategy extends RotationStrategy {
+  rotateClockwise() {
+
+  }
+
+  rotateCounterClockwise() {
+
+  }
+}
+
+
+export class IPieceRotationStrategy extends RotationStrategy {
+  rotateClockwise() {
+
+  }
+
+  rotateCounterClockwise() {
+
+  }
+}
+
+
+export class TSZJLPieceRotationStrategy extends RotationStrategy {
+  rotateCounterClockwise() {
+    const rotatedCoordinates = this.piece.coords.map(
+      coord => {
+        const transposedCoord = new GridCoordinate({ col: coord.col - this.piece.pivot.col, row: coord.row - this.piece.pivot.row });
+        const rotatedCoord = new GridCoordinate({ col: transposedCoord.row, row: - transposedCoord.col });
+        return new GridCoordinate({ col: rotatedCoord.col + this.piece.pivot.col, row: rotatedCoord.row + this.piece.pivot.row })
+      });
+
+    // go through all 5 transformations to see
+    this.piece.coords = rotatedCoordinates;
+  }
+
+  rotateClockwise() {
+    const rotatedCoordinates = this.piece.coords.map(
+      coord => {
+        const transposedCoord = new GridCoordinate({ col: coord.col - this.piece.pivot.col, row: coord.row - this.piece.pivot.row });
+        const rotatedCoord = new GridCoordinate({ col: - transposedCoord.row, row: transposedCoord.col });
+        return new GridCoordinate({ col: rotatedCoord.col + this.piece.pivot.col, row: rotatedCoord.row + this.piece.pivot.row })
+      });
+
+    // go through all 5 transformations to see
+    this.piece.coords = rotatedCoordinates;
+  }
+}
